Change of model generation approach
Instead of managing the model and data by "model.py", we gave this management to "model_generator.py".
The reason for this change is that we only have one model type in this project, so we don't need to create a redundant layer of abstraction.
"model.py" only has some functionalities to serve to "model_generator"
Minor changes